finish splitting up heatmap
